Cohort builder: shared criteria event card + several UX fixes
Refactor the cohort builder's per-event cards onto one shared component
and fix a batch of issues found while testing develop.

Shared criteria event card
- Add CriteriaEventCard: renders a single event's card (type label,
  concept-set picker, add-attribute, delete; optional cardinality sidebar
  and temporal-window chip; recursive nested criteria).
- Move the concept-set selector into the card header everywhere.
- Entry events, CriteriaGroupEditor and NestedCriteriaEditor all render
  CriteriaEventCard, so nested criteria now match the normal cards. Remove
  the old EntryEventCard and the divergent nested-card markup.
- Add a compact mode to EventConceptSetField for header use.

Cohort entry events
- Continuous-observation editor: replace the page-dimming modal with an
  anchored popover and rename the pill to name the concept (discussion #99).
- Remove the non-editable "At least 1" cardinality sidebar from entry events.
- New concept sets open on the Search tab, existing ones on Selected
  (discussion #97).

Concept detail
- Handle "Add to Concept Set" from a selected concept (drawer + full page).
- Grey out stale content while a different concept loads.
- Open related/hierarchy concepts in the side panel instead of navigating
  the whole app.

Config / data sources
- Fix data-source deletion (use the numeric sourceId, not sourceKey).
- Keep the permissions tab bar visible when the user/permission list is long.

Tests added/updated accordingly; type-check and lint are clean.
