+++ src/locales/en.json
@@ -447,6 +447,11 @@
       "cohortEntryEventsText_5": "days after event index date",
       "cohortEntryEventsText_6": "Limit initial events to:",
       "cohortEntryEventsText_7": "per person.",
+      "continuousObservationLabel": "Continuous observation",
+      "continuousObservationTitle": "Continuous observation window",
+      "continuousObservationBefore": "Days before",
+      "continuousObservationAfter": "Days after",
+      "continuousObservationHelp": "People without this much continuous observation around the index date are excluded from the cohort.",
       "cohortErasText_1": "Specify era collapse gap size:",
       "cohortErasText_2": "Left censor cohort start dates to",
       "cohortErasText_2_1": " ",
