+++ tests/component/concepts/detail/ConceptHierarchyMiniMap.spec.ts
@@ -1,4 +1,4 @@
-import { describe, it, expect, beforeEach } from 'vitest'
+import { describe, it, expect, beforeEach, vi } from 'vitest'
 import { mount } from '@vue/test-utils'
 import { createVuetify } from 'vuetify'
 import * as components from 'vuetify/components'
@@ -6,6 +6,7 @@ import * as directives from 'vuetify/directives'
 import { createPinia, setActivePinia } from 'pinia'
 import { createRouter, createMemoryHistory } from 'vue-router'
 import ConceptHierarchyMiniMap from '@/components/concepts/detail/ConceptHierarchyMiniMap.vue'
+import { useConceptDetailDrawerStore } from '@/stores/concept-detail-drawer'
 import type { Concept } from '@/models/concept-set.types'
 import type { RelatedConcept } from '@/models/concept-detail.types'
 
@@ -72,6 +73,23 @@ describe('ConceptHierarchyMiniMap', () => {
     const current = wrapper.find('[data-testid="hierarchy-current"]')
     expect(current.exists()).toBe(true)
     expect(current.text()).toContain('Type 2 diabetes mellitus')
+  })
+
+  it('opens a parent/child concept in the side-panel drawer instead of routing', async () => {
+    const vuetify = createVuetify({ components, directives })
+    const router = createRouter({ history: createMemoryHistory(), routes: [] })
+
+    const wrapper = mount(ConceptHierarchyMiniMap, {
+      props: { concept, parents, children, sourceKey: 'OHDSI' },
+      global: { plugins: [vuetify, router] },
+    })
+
+    const drawer = useConceptDetailDrawerStore()
+    const openSpy = vi.spyOn(drawer, 'open')
+
+    // First clickable node link is the parent concept.
+    await wrapper.find('a.node-link').trigger('click')
+    expect(openSpy).toHaveBeenCalledWith('OHDSI', 73211009)
   })
 
   it('shows empty placeholder when no parents and no children', () => {
